requirements.txt
streamlit
pinecone
sentence-transformers
python-dotenv
pandas
requests